Update Main.java
Fixed an issue with getting collumn number where it would not subtract 1 for double digit collumns (thx for letting me know Eduardo)

slot class added as a middle man between vending machine class and item class

jasmine look at the vending machine constructor to see where you need to implement the database.

added in dev option to manually add an item. items are predefined in the vending machine class.
currently stackItem() method works with this dev option. in addition after manually adding an item into the vending machine, customerI() will recognize that an item is in the slot code where you manually entered the item. working on displaySlots() is now possible with this dev option as a way to test